fix express tests
