+++ express-api/src/routes/usersRouter.ts
@@ -31,21 +31,21 @@ router.route(`/access/requests`).post(catchErrors(submitUserAccessRequest));
 router.route(`/agencies/:username`).get(catchErrors(getUserAgencies));
 
 router
-  .route(`/users`)
+  .route(`/`)
   .get(catchErrors(getUsers))
   .post(protectedRoute([Roles.ADMIN]), catchErrors(addUser));
 
-router.route(`/users/my/agency`).post(catchErrors(getUsersSameAgency)); // TODO: Should this just be generic: get users from an agency?
+router.route(`/my/agency`).post(catchErrors(getUsersSameAgency)); // TODO: Should this just be generic: get users from an agency?
 
-router.route(`/users/roles`).get(catchErrors(getAllRoles));
+router.route(`/roles`).get(catchErrors(getAllRoles));
 
 router
-  .route(`/users/roles/:username`)
+  .route(`/roles/:username`)
   .get(catchErrors(getUserRolesByName))
   .put(protectedRoute([Roles.ADMIN]), catchErrors(updateUserRolesByName));
 
 router
-  .route(`/users/:id`)
+  .route(`/:id`)
   .get(catchErrors(getUserById))
   .put(protectedRoute([Roles.ADMIN]), catchErrors(updateUserById)) // TODO: should put be a patch?
   .delete(protectedRoute([Roles.ADMIN]), deleteUserById);
