+++ express-api/tests/integration/administrativeAreas/administrativeAreas.test.ts
@@ -6,7 +6,7 @@ import { UUID } from 'crypto';
 
 const request = supertest(app);
 
-const ADMIN_AREAS_PATH = '/v2/admin/administrativeAreas';
+const ADMIN_AREAS_PATH = '/v2/administrativeAreas';
 const mockAdministrativeArea: IAdministrativeArea = {
   createdOn: faker.date.anytime().toLocaleString(),
   updatedOn: faker.date.anytime().toLocaleString(),
