--- express-api/tests/unit/controllers/admin/users/usersController.test.ts
@@ -1,321 +1,0 @@
-import { Request, Response } from 'express';
-import controllers from '@/controllers';
-import {
-  MockReq,
-  MockRes,
-  getRequestHandlerMocks,
-  produceUser,
-} from '../../../../testUtils/factories';
-import { Roles } from '@/constants/roles';
-import { faker } from '@faker-js/faker';
-import { UUID } from 'crypto';
-import { updateUserRolesByName } from '@/controllers/admin/users/usersController';
-
-let mockRequest: Request & MockReq, mockResponse: Response & MockRes;
-
-const {
-  addUser,
-  //addUserRoleByName,
-  getUserById,
-  getUserRolesByName,
-  getAllRoles,
-  getUsers,
-  //getUsersSameAgency,
-  deleteUserById,
-  //deleteUserRoleByName,
-  updateUserById,
-} = controllers.admin;
-
-const _getUsers = jest.fn().mockImplementation(() => {
-  return [produceUser()];
-});
-const _addUser = jest.fn().mockImplementation((user) => {
-  return user;
-});
-const _updateUser = jest.fn().mockImplementation((user) => {
-  return user;
-});
-const _deleteUser = jest.fn().mockImplementation((user) => {
-  return user;
-});
-const _getRoles = jest.fn().mockImplementation(() => {
-  return ['admin', 'test'];
-});
-const _getUserRoles = jest.fn().mockImplementation(() => {
-  return ['admin'];
-});
-const _updateUserRoles = jest.fn().mockImplementation((username, roles) => {
-  return roles;
-});
-const _getUserById = jest.fn().mockImplementation(() => produceUser());
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-jest.mock('@/services/admin/usersServices', () => ({
-  getUsers: () => _getUsers(),
-  addUser: () => _addUser(),
-  updateUser: () => _updateUser(),
-  deleteUser: () => _deleteUser(),
-  getKeycloakRoles: () => _getRoles(),
-  getKeycloakUserRoles: () => _getUserRoles(),
-  updateKeycloakUserRoles: () => _updateUserRoles(),
-  getUserById: () => _getUserById(),
-}));
-
-describe('UNIT - Users Admin', () => {
-  beforeEach(() => {
-    const { mockReq, mockRes } = getRequestHandlerMocks();
-    mockRequest = mockReq;
-    mockRequest.setUser({ client_roles: [Roles.ADMIN] });
-    mockResponse = mockRes;
-  });
-
-  describe('Controller getUsers', () => {
-    it('should return status 200 and a list of users', async () => {
-      await getUsers(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      expect(Array.isArray(mockResponse.sendValue)).toBe(true);
-    });
-
-    it('should return a list of users based off the filter', async () => {
-      mockRequest.query = {
-        position: 'Tester',
-      };
-      await getUsers(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      expect(Array.isArray(mockResponse.sendValue)).toBe(true);
-      expect(mockResponse.sendValue.length === 1);
-    });
-
-    it('should return status 400 when given a filter that fails to parse', async () => {
-      mockRequest.query = {
-        page: 'hi',
-      };
-      await getUsers(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-      expect(mockResponse.sendValue).toBe('Failed to parse filter query.');
-    });
-  });
-
-  // describe('Controller getUsersSameAgency', () => {
-  //   // TODO: remove stub test when controller is complete
-  //   it('should return the stub response of 501', async () => {
-  //     await getUsersSameAgency(mockRequest, mockResponse);
-  //     expect(mockResponse.statusValue).toBe(501);
-  //   });
-
-  //   // TODO: enable other tests when controller is complete
-  //   xit('should return status 200 and a list of users', async () => {
-  //     await getUsersSameAgency(mockRequest, mockResponse);
-  //     expect(mockResponse.statusValue).toBe(200);
-  //   });
-  // });
-
-  describe('Controller addUser', () => {
-    const user = produceUser();
-    beforeEach(() => {
-      _addUser.mockImplementation(() => user);
-      mockRequest.body = user;
-    });
-
-    it('should return status 201 and the new user', async () => {
-      await addUser(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(201);
-      expect(mockResponse.sendValue.Id).toBe(mockRequest.body.Id);
-    });
-
-    it('should return status 400 when the userService.addUser throws an error', async () => {
-      _addUser.mockImplementationOnce(() => {
-        throw new Error();
-      });
-      await addUser(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-    });
-  });
-
-  describe('Controller getUserById', () => {
-    const user = produceUser();
-    beforeEach(() => {
-      _getUserById.mockImplementation(() => user);
-    });
-
-    it('should return status 200 and the user info', async () => {
-      mockRequest.params.id = user.Id;
-      await getUserById(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      expect(mockResponse.sendValue.Id).toBe(user.Id);
-    });
-
-    it('should return status 400 if the uuid cannot be parsed', async () => {
-      mockRequest.params.id = 'hello';
-      await getUserById(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-      expect(mockResponse.sendValue).toBe('Could not parse UUID.');
-    });
-
-    it('should return status 404 userService.getUserById does not find a user', async () => {
-      mockRequest.params.id = user.Id;
-      _getUserById.mockImplementationOnce(() => undefined);
-      await getUserById(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(404);
-    });
-  });
-
-  describe('Controller updateUserById', () => {
-    const user = produceUser();
-    user.Email = 'newEmail@gov.bc.ca';
-    beforeEach(() => {
-      _updateUser.mockImplementation(() => user);
-      mockRequest.params.id = user.Id;
-      mockRequest.body = { ...user, Email: 'newEmail@gov.bc.ca' };
-    });
-
-    it('should return status 200 and the updated user', async () => {
-      await updateUserById(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      expect(mockResponse.sendValue.Email).toBe('newEmail@gov.bc.ca');
-    });
-
-    it('should return status 400 if the param ID does not match the body ID', async () => {
-      mockRequest.params.id = faker.string.uuid() as UUID;
-      await updateUserById(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-      expect(mockResponse.sendValue).toBe('The param ID does not match the request body.');
-    });
-
-    it('should throw an error if userService.updateUser throws an error', async () => {
-      _updateUser.mockImplementation(() => {
-        throw new Error();
-      });
-      expect(async () => await updateUserById(mockRequest, mockResponse)).rejects.toThrow();
-    });
-  });
-
-  describe('Controller deleteUserById', () => {
-    const user = produceUser();
-    beforeEach(() => {
-      _deleteUser.mockImplementation(() => user);
-      mockRequest.params.id = user.Id;
-      mockRequest.body = user;
-    });
-
-    it('should return status 200', async () => {
-      await deleteUserById(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      expect(mockResponse.sendValue.Id).toBe(user.Id);
-    });
-
-    it('should return status 400 if the param ID does not match the body ID', async () => {
-      mockRequest.params.id = faker.string.uuid() as UUID;
-      await deleteUserById(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-      expect(mockResponse.sendValue).toBe('The param ID does not match the request body.');
-    });
-
-    it('should throw an error if userService.deleteUser throws an error', async () => {
-      _deleteUser.mockImplementationOnce(() => {
-        throw new Error();
-      });
-      expect(async () => await deleteUserById(mockRequest, mockResponse)).rejects.toThrow();
-    });
-  });
-
-  describe('Controller getUserRolesByName', () => {
-    beforeEach(() => {
-      mockRequest.params.username = 'test';
-    });
-
-    it('should return status 200 and a list of their roles', async () => {
-      await getUserRolesByName(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      // Only role is Admin
-      expect(mockResponse.sendValue).toHaveLength(1);
-      expect(mockResponse.sendValue.at(0)).toBe('admin');
-    });
-
-    it('should return status 400 if params.username is not provided', async () => {
-      mockRequest.params = {};
-      await getUserRolesByName(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-      expect(mockResponse.sendValue).toBe('Username was empty.');
-    });
-  });
-
-  describe('Controller getAllRoles', () => {
-    it('should return status 200 and a list of roles assigned to a user', async () => {
-      await getAllRoles(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      expect(mockResponse.sendValue.at(0)).toBe('admin');
-    });
-
-    it('should throw an error when internal service throws', async () => {
-      _getRoles.mockImplementationOnce(() => {
-        throw Error();
-      });
-      expect(async () => await getAllRoles(mockRequest, mockResponse)).rejects.toThrow();
-    });
-  });
-
-  describe('Controller updateUserRolesByName', () => {
-    const user = produceUser();
-    beforeEach(() => {
-      _updateUserRoles.mockImplementation((username, roles) => roles);
-      mockRequest.params = {
-        username: user.Username,
-      };
-      mockRequest.body = ['admin', 'auditor'];
-    });
-    it('should return status 200 and a list of updated roles', async () => {
-      await updateUserRolesByName(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(200);
-      // TODO: Check the return value. It's currently undefined for some reason.
-    });
-
-    it('should return status 400 if the request body was not parsed successfully', async () => {
-      mockRequest.body = {
-        notGood: true,
-      };
-      await updateUserRolesByName(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-      expect(mockResponse.sendValue).toBe('Request body was wrong format.');
-    });
-
-    it('should return 400 if params.username was not provided', async () => {
-      mockRequest.params = {};
-      await updateUserRolesByName(mockRequest, mockResponse);
-      expect(mockResponse.statusValue).toBe(400);
-      expect(mockResponse.sendValue).toBe('Username was empty.');
-    });
-  });
-
-  // describe('Controller addUserRoleByName', () => {
-  //   beforeEach(() => {
-  //     mockRequest.params.username = `test`;
-  //     mockRequest.body = 'new role';
-  //   });
-
-  //   it('should return status 201 and the updated user with that role', async () => {
-  //     await addUserRoleByName(mockRequest, mockResponse);
-  //     expect(mockResponse.statusValue).toBe(201);
-  //     expect(mockResponse.sendValue).toContain('new role');
-  //   });
-  // });
-
-  // describe('Controller deleteUserRoleByName', () => {
-  //   beforeEach(() => {
-  //     mockRequest.params.username = `${mockUser.Username}`;
-  //     mockRequest.body = 'new role';
-  //   });
-  //   // TODO: remove stub test when controller is complete
-  //   it('should return the stub response of 501', async () => {
-  //     await deleteUserRoleByName(mockRequest, mockResponse);
-  //     expect(mockResponse.statusValue).toBe(501);
-  //   });
-
-  //   // TODO: enable other tests when controller is complete
-  //   xit('should return status 204 and updated user without role', async () => {
-  //     await deleteUserRoleByName(mockRequest, mockResponse);
-  //     expect(mockResponse.statusValue).toBe(204);
-  //     expect(mockResponse.jsonValue.roles).not.toContain('new role');
-  //   });
-  // });
-});
